deletar pedido

+++ src/main/java/com/comandago/api/services/PedidoService.java
@@ -90,8 +90,18 @@ public class PedidoService {
         }
     }
 
-    public void excluirPedido(Long id) {
-        pedidoRepository.deleteById(id);
+    public boolean excluirPedido(Long id) {
+        Optional<Pedido> pedidOptional = pedidoRepository.findById(id);
+        if(pedidOptional.isPresent()){
+            Pedido pedido = pedidOptional.get();
+            Optional<Comanda> comandOptional = comandaRepository.findById(pedido.getIdComanda());
+            Comanda comanda = comandOptional.get();
+            comanda.setTotalAPagar(comanda.getTotalAPagar()-pedido.getValor());
+            comandaRepository.save(comanda);
+            pedidoRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
     public boolean adicionarItens(Long idPedido, List<ItemPedidoDTO> itens){
